Updated ximu libraries

+++ backend/Arduino/ping_client/XimuReceiver.h
@@ -76,9 +76,9 @@ typedef struct {
 
 typedef struct {
 	// Strength: 1 ~ 16, as indications of strength
-	short strength;
+	long strength;
 	// ID of drum: 1 ~ 7
-	unsigned short drumID;
+	unsigned long drumID;
 } DrumSet;
 
 //------------------------------------------------------------------------------
